@@ -27,7 +27,7 @@ import Mustache
 class TemplateRepositoryPathTests: XCTestCase {
     
     func testTemplateRepositoryWithURL() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         let directoryPath = testBundle.path(forResource: "TemplateRepositoryFileSystemTests_UTF8", ofType: nil)!
         let repo = TemplateRepository(directoryPath: directoryPath)
         var template: Template
@@ -57,7 +57,7 @@ class TemplateRepositoryPathTests: XCTestCase {
     }
     
     func testTemplateRepositoryWithURLTemplateExtensionEncoding() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         var directoryPath: String
         var repo: TemplateRepository
         var template: Template
@@ -101,7 +101,7 @@ class TemplateRepositoryPathTests: XCTestCase {
     }
     
     func testAbsolutePartialName() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         let directoryPath = testBundle.path(forResource: "TemplateRepositoryFileSystemTests", ofType: nil)!
         let repo = TemplateRepository(directoryPath: directoryPath)
         let template = try! repo.template(named: "base")
@@ -110,7 +110,7 @@ class TemplateRepositoryPathTests: XCTestCase {
     }
     
     func testPartialNameCanNotEscapeTemplateRepositoryRootDirectory() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         let directoryPath = testBundle.path(forResource: "TemplateRepositoryFileSystemTests", ofType: nil)!
         let repo = TemplateRepository(directoryPath: (directoryPath as NSString).appendingPathComponent("partials"))
         
@@ -126,5 +126,23 @@ class TemplateRepositoryPathTests: XCTestCase {
         } catch {
             XCTFail("Expected MustacheError")
         }
+    }
+
+    func testFileNameContainsWhitespace() {
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
+        let directoryPath = testBundle.path(forResource: "TemplateRepositoryFileSystemTests", ofType: nil)!
+        let repo = TemplateRepository(directoryPath: directoryPath)
+        let template = try! repo.template(named: "contains whitespace")
+        let rendering = try! template.render()
+        XCTAssertEqual(rendering, "success\n")
+    }
+
+    func testFileAndDictonaryNameContainsWhitespace() {
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
+        let directoryPath = testBundle.path(forResource: "TemplateRepositoryFileSystemTests/contains whitespace", ofType: nil)!
+        let repo = TemplateRepository(directoryPath: directoryPath)
+        let template = try! repo.template(named: "contains whitespace")
+        let rendering = try! template.render()
+        XCTAssertEqual(rendering, "success\n")
     }
 }
