@@ -27,7 +27,7 @@ import Mustache
 class TemplateRepositoryURLTests: XCTestCase {
     
     func testTemplateRepositoryWithURL() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         let URL = testBundle.url(forResource: "TemplateRepositoryFileSystemTests_UTF8", withExtension: nil)!
         let repo = TemplateRepository(baseURL: URL)
         var template: Template
@@ -57,7 +57,7 @@ class TemplateRepositoryURLTests: XCTestCase {
     }
     
     func testTemplateRepositoryWithURLTemplateExtensionEncoding() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         var URL: Foundation.URL
         var repo: TemplateRepository
         var template: Template
@@ -101,7 +101,7 @@ class TemplateRepositoryURLTests: XCTestCase {
     }
     
     func testAbsolutePartialName() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         let URL = testBundle.url(forResource: "TemplateRepositoryFileSystemTests", withExtension: nil)!
         let repo = TemplateRepository(baseURL: URL)
         let template = try! repo.template(named: "base")
@@ -110,7 +110,7 @@ class TemplateRepositoryURLTests: XCTestCase {
     }
     
     func testPartialNameCanNotEscapeTemplateRepositoryRootURL() {
-        let testBundle = Bundle(for: type(of: self))
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
         let URL = testBundle.url(forResource: "TemplateRepositoryFileSystemTests", withExtension: nil)!
         let repo = TemplateRepository(baseURL: URL.appendingPathComponent("partials"))
         
@@ -126,5 +126,23 @@ class TemplateRepositoryURLTests: XCTestCase {
         } catch {
             XCTFail("Expected MustacheError")
         }
+    }
+
+    func testFileNameContainsWhitespace() {
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
+        let directoryPath = testBundle.url(forResource: "TemplateRepositoryFileSystemTests", withExtension: nil)!
+        let repo = TemplateRepository(baseURL: directoryPath)
+        let template = try! repo.template(named: "contains whitespace")
+        let rendering = try! template.render()
+        XCTAssertEqual(rendering, "success\n")
+    }
+
+    func testFileAndDictionaryNameContainsWhitespace() {
+        let testBundle = Bundle.ofFileDirectory(filePath: #file)!
+        let directoryPath = testBundle.url(forResource: "TemplateRepositoryFileSystemTests/contains whitespace", withExtension: nil)!
+        let repo = TemplateRepository(baseURL: directoryPath)
+        let template = try! repo.template(named: "contains whitespace")
+        let rendering = try! template.render()
+        XCTAssertEqual(rendering, "success\n")
     }
 }
